==> notebooks/02_faster_rcnn_resnet50.ipynb <==
import sys
sys.path.append('..')

import json
from pathlib import Path
from textwrap import dedent
import pdb

from tqdm import tqdm
from ipywidgets import interact
from PIL import Image, ImageDraw, ImageFont
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from matplotlib import patches, patheffects

import torch
import torchvision
from torchvision.models.detection.faster_rcnn import FastRCNNPredictor

import pylib.transforms as T
from pylib.engine import train_one_epoch, evaluate
import pylib.utils as utils

from pylib.label_babel_dataset import LabelBabelDataset

%matplotlib inline
%reload_ext autoreload
%autoreload 2

%load_ext watermark
%watermark -i -z -v -p torch,torchvision

DATA_DIR = Path('..') / 'data'
VAL_DIR = DATA_DIR / 'valid'
TRN_DIR = DATA_DIR / 'train'
MODEL_DIR = DATA_DIR / 'models'

TRN_CSV = DATA_DIR / 'train.csv'
VAL_CSV = DATA_DIR / 'valid.csv'

BOX = 'box'
CAT = 'category'
PATH = 'path'
CLASS = 'class'
SUB_ID = 'subject_id'

SEED = 23

DEVICE = torch.device('cuda')
torch.backends.cudnn.benchmark = True  # Optimizes cudnn

CATS = ['background', 'handwritten', 'typewritten']
CLASSES = len(CATS)

EPOCHS = 10  # 50
CHECKPOINT = 'checkpoint_{}.pth.tar'

torch.manual_seed(SEED)

trn_df = pd.read_csv(TRN_CSV, index_col='subject_id').reset_index()
val_df = pd.read_csv(VAL_CSV, index_col='subject_id').reset_index()

trn_df.head()

model = torchvision.models.detection.fasterrcnn_resnet50_fpn(pretrained=True)

in_features = model.roi_heads.box_predictor.cls_score.in_features

model.roi_heads.box_predictor = FastRCNNPredictor(in_features, CLASSES)

def get_transform(train):
    transforms = [T.ToTensor()]
    if train:
        transforms.append(T.RandomHorizontalFlip(0.5))
        transforms.append(T.RandomVerticalFlip(0.5))
        transforms.append(T.RandomRotate())
        transforms.append(T.ColorJitter(
            brightness=0.25, contrast=0.25, saturation=0.25, hue=0.25))
    return T.Compose(transforms)

trn_dataset = LabelBabelDataset(trn_df, get_transform(train=True))
val_dataset = LabelBabelDataset(val_df, get_transform(train=False))

trn_loader = torch.utils.data.DataLoader(
    trn_dataset, batch_size=1, shuffle=True, num_workers=4,
    collate_fn=utils.collate_fn)

val_loader = torch.utils.data.DataLoader(
    val_dataset, batch_size=1, shuffle=False, num_workers=4,
    collate_fn=utils.collate_fn)

model.to(DEVICE)

params = [p for p in model.parameters() if p.requires_grad]
optimizer = torch.optim.SGD(
    params, lr=0.005, momentum=0.9, weight_decay=0.0005)

lr_scheduler = torch.optim.lr_scheduler.StepLR(
    optimizer, step_size=5, gamma=0.1)

last_checkpoint = sorted(MODEL_DIR.glob(CHECKPOINT.format('*')))[-1]

state = torch.load(last_checkpoint)

first_epoch = state['epoch']
model.load_state_dict(state['state_dict'])
optimizer.load_state_dict(state['optimizer'])

last_checkpoint

def save_state(epoch, model, optimizer):
    state_path = MODEL_DIR / CHECKPOINT.format(str(epoch).zfill(3))
    if state_path.exists():
        return
    state = {
        'epoch': epoch + 1,
        'state_dict': model.state_dict(),
        'optimizer': optimizer.state_dict(),
    }
    torch.save(state, state_path)

def train_model():
    for epoch in range(first_epoch, EPOCHS):
        train_one_epoch(
            model, optimizer, trn_loader, DEVICE, epoch, print_freq=100)
        lr_scheduler.step()
        evaluate(model, val_loader, device=DEVICE)
        if state % 10 == 0:
            save_state(epoch, model, optimizer)

    save_state(epoch, model, optimizer)


# train_model()
