==> notebooks/04_pipeline.ipynb <==
from pathlib import Path
from shutil import rmtree
import pdb

from tqdm import tqdm
from PIL import Image, ImageDraw, ImageFont
import numpy as np
import pandas as pd

import torch
import torchvision
from torchvision.models.detection.faster_rcnn import FastRCNNPredictor

import transforms as T
import utils

from label_babel_dataset import LabelBabelDataset

DATA_DIR = Path('.') / 'data'

TRN_DIR = DATA_DIR / 'train'
VAL_DIR = DATA_DIR / 'valid'
MODEL_DIR = DATA_DIR / 'models'
HAND_DIR = DATA_DIR / 'handwritten'
TYPE_DIR = DATA_DIR / 'typewritten'

VAL_CSV = DATA_DIR / 'valid.csv'
TRN_CSV = DATA_DIR / 'train.csv'

BOX = 'box'
CAT = 'category'
ORI = 'original'
PATH = 'path'
CROP = 'predicted_box_original'
CLASS = 'class'
SUB_ID = 'subject_id'
PRED_BOX = 'predicted_box'
PRED_CAT = 'predicted_category'
PRED_CLASS = 'predicted_class'

DEVICE = torch.device('cuda')
torch.backends.cudnn.benchmark = True

HANDWRITTEN = 'handwritten'
TYPEWRITTEN = 'typewritten'
CATS = ['background', HANDWRITTEN, TYPEWRITTEN]
CLASSES = len(CATS)

CHECKPOINT = 'checkpoint_{}.pth.tar'

SCALE = 4.0

def load_df(df_path):
    df = pd.read_csv(df_path, index_col='subject_id').reset_index()
    df[PRED_CLASS] = None
    df[PRED_CAT] = None
    df[PRED_BOX] = None
    df[CROP] = None
    return df  # df.iloc[:2, :].copy()

val_df = load_df(VAL_CSV)
trn_df = load_df(TRN_CSV)

model = torchvision.models.detection.fasterrcnn_resnet50_fpn(pretrained=True)

in_features = model.roi_heads.box_predictor.cls_score.in_features

model.roi_heads.box_predictor = FastRCNNPredictor(in_features, CLASSES)

transforms = T.Compose([T.ToTensor()])

val_dataset = LabelBabelDataset(val_df, transforms)
trn_dataset = LabelBabelDataset(trn_df, transforms)

model.to(DEVICE)
model.eval();

last_checkpoint = sorted(MODEL_DIR.glob(CHECKPOINT.format('*')))[-1]

state = torch.load(last_checkpoint)

model.load_state_dict(state['state_dict'])

state['epoch']

def get_predictions(dataset, df):
    for idx, (image, target) in tqdm(enumerate(dataset)):
        with torch.no_grad():
            prediction = model([image.to(DEVICE)])

        p_class = 0
        p_box = [0, 0, 0, 0]
        if len(prediction[0]['labels'].cpu()):
            p_class = prediction[0]['labels'].cpu()[0].numpy()
            p_box = prediction[0]['boxes'].cpu()[0].numpy()

        df.at[idx, PRED_BOX] = p_box
        df.at[idx, PRED_CLASS] = p_class


get_predictions(val_dataset, val_df)
get_predictions(trn_dataset, trn_df)

cats = pd.Series(CATS)

trn_df[PRED_CAT] = trn_df[PRED_CLASS].astype(int).map(cats)
val_df[PRED_CAT] = val_df[PRED_CLASS].astype(int).map(cats)

val_df.head()

rmtree(HAND_DIR, ignore_errors=True)
rmtree(TYPE_DIR, ignore_errors=True)

HAND_DIR.mkdir()
TYPE_DIR.mkdir()

def crop_image(df):
    for idx, row in tqdm(df.iterrows()):
        path = Path(row.at[ORI])
        src = str(path)
        dst = TYPE_DIR if row.at[PRED_CAT] == TYPEWRITTEN else HAND_DIR
        dst = dst / path.name
        box = [int(round(x * SCALE)) for x in row.at[PRED_BOX]]
        df.at[idx, CROP] = box
        if sum(x for x in box) == 0:
            continue
        image = Image.open(src).convert('RGB')
        width, height = image.size
        image = image.crop(box)
        if width > height:
            image = image.transpose(Image.ROTATE_270)
        image.save(dst)


crop_image(val_df)
crop_image(trn_df)

val_df.to_csv(VAL_CSV, index=False)
trn_df.to_csv(TRN_CSV, index=False)
