==> notebooks/06_copy_gold.ipynb <==
from pathlib import Path
from shutil import copyfile, rmtree
import pandas as pd

DATA_DIR = Path('.') / 'data'

GOLD_LAB_DIR = DATA_DIR / 'gold_labels'
GOLD_RAW_DIR = DATA_DIR / 'gold_raw_data'

RAW_DIR = DATA_DIR / 'images'
TYPEWRITTEN_DIR = DATA_DIR / 'typewritten'

GOLD_STD = DATA_DIR / 'GoldStandard_type100.csv'
GOLD_STD = DATA_DIR / 'GoldStandardExtraction_NewOCR200.csv'

SUB_ID = 'subject_id'

rmtree(GOLD_LAB_DIR, ignore_errors=True)
rmtree(GOLD_RAW_DIR, ignore_errors=True)

GOLD_LAB_DIR.mkdir()
GOLD_RAW_DIR.mkdir()

df = pd.read_csv(GOLD_STD)
df.head()

for subject_id in df.subject_id:
    print(subject_id)
    file_name = str(int(subject_id)) + '.jpg'

    src = RAW_DIR / file_name
    dst = GOLD_RAW_DIR / file_name
    copyfile(src, dst)

    src = TYPEWRITTEN_DIR / file_name
    dst = GOLD_LAB_DIR / file_name
    copyfile(src, dst)
