==> notebooks/07_score_ocr.ipynb <==
import sys
sys.path.append('..')

from pathlib import Path
import regex
import pandas as pd

DATA_DIR = Path('..') / 'data'
CORPORA = DATA_DIR / 'corpora'

WORDS_EN = CORPORA / 'nltk' / 'en'
MASC_500 = CORPORA / 'masc_500k_texts'

US_STATES = CORPORA / 'traiter' / 'US_states.csv'
US_COUNTIES = CORPORA / 'traiter' / 'US_counties.csv'
NAME_PARTS = CORPORA / 'traiter' / 'name_parts.csv'

ITIS_TAXA = CORPORA / 'plant_taxa.csv'
CITIES = CORPORA / 'US' / 'US.txt'

OCR_CSV = DATA_DIR / 'OCR_IE-partII_score.csv'

SPLITTER = '[^a-z]+'
VOWELS = '[aeiouy]'
CHARS = r'[\u0020-\u007e\s]'

def strip_words(words):
    words = {w for w in words
             if len(w) == 1 or regex.search(VOWELS, w)}
    return words

corpora = set()

with open(WORDS_EN) as in_file:
    words = in_file.read().lower().split()
    corpora |= strip_words(words)

for path in MASC_500.glob('**/*.txt'):
    with open(path, encoding='ISO-8859-1') as text_file:
        text = text_file.read()
    words = set(regex.split(SPLITTER, text.lower()))
    corpora |= strip_words(words)

df = pd.read_csv(US_STATES, na_filter=False)
corpora |= set(df['postal'].str.lower())

words = df['state'].str.lower().str.split()
words = {regex.sub(SPLITTER, '', a) for aa in words for a in aa if a}

words = df['abbrev'].str.lower().str.split(',')
words = {regex.sub(SPLITTER, '', a) for aa in words for a in aa if a}
corpora |= strip_words(words)

df = pd.read_csv(US_COUNTIES, na_filter=False)
words = df['County'].str.lower().str.split(SPLITTER)
words = {a for aa in words for a in aa if a}
corpora |= strip_words(words)

df = pd.read_csv(NAME_PARTS, na_filter=False)
words = set(df['name'].str.lower())
corpora |= strip_words(words)

df = pd.read_csv(ITIS_TAXA, na_filter=False)
words = df['complete_name'].str.lower().str.split(SPLITTER)
words = {a for aa in words for a in aa if a}
corpora |= strip_words(words)

df = pd.read_csv(CITIES, sep='\t', header=None)
words = df[2].str.lower().str.split(SPLITTER)
words = {a for aa in words for a in aa if a}
corpora |= strip_words(words)

len(corpora)

word_scores = []
char_scores = []

df = pd.read_csv(OCR_CSV, na_filter=False, index_col='subject_id')
for subject_id, row in df.iterrows():
    text = row['predicted_text'].lower()

    chars = list(text)
    found = sum(1 for c in chars if regex.match(CHARS, c))
    char_score = round(found / len(chars), 4)
    char_scores.append(char_score)

    words = regex.split(SPLITTER, text.lower())
    found = sum(1 for w in words if w in corpora)
    word_score = round(found / len(words), 4)
    word_scores.append(word_score)
    # print(f'{subject_id:<10}\t{word_score:0.4f}\t{char_score:0.4f}')

data = {
    'subject_id': df.index,
    'word_score': word_scores,
    'char_score': char_scores}

df2 = pd.DataFrame(data=data)
# df2.to_csv(DATA_DIR / 'OCR_scores.csv', index=False)
df2.head()

text = df.at[10668179, 'predicted_text']
chars = list(text)
for char in chars:
    print(f'{char} {regex.match(CHARS, char) is not None}')
#  = sum(1 for c in chars if regex.match(CHARS, c))
